# 00_docs/openapi_api_contract104.yaml
openapi: 3.0.3
info:
  title: MGF — Modular Generation Framework API
  description: |
    REST API for the MGF platform. All endpoints are prefixed with `/v1`.

    ## Authentication
    All protected routes require a Bearer token obtained from `POST /auth/login`.
    Pass it as: `Authorization: Bearer <token>`

    ## AI Generation — per-user keys
    Every user brings their own AI keys (Anthropic, OpenAI, Gemini, LM Studio,
    or a custom MiniMax-style proxy). Keys are stored encrypted in
    `user_ai_providers` and decrypted server-side per request — they are
    **never** sent to the browser.

    Generation calls reference a specific `provider_id` (the user's saved row)
    so the backend knows whose key, base URL, and default model to use.

    Pagination, error formats, and other conventions are unchanged from v1.0.3.

    ## Pagination
    List endpoints support `?page=1&per_page=20`. Response includes a `meta` pagination object.

    ## Error Format
    All errors return `{ "message": "...", "errors": { "field": ["..."] } }` (422)
    or `{ "error": "...", "code": "..." }` for typed AI errors (404 / 422 / 502).
  version: 1.0.4
  contact:
    name: MGF Team
servers:
  - url: http://mgf.vortex-tech.tech/api/v1/
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development

# ─────────────────────────────────────────────────────────────────────────────
# SECURITY
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# TAGS (module grouping)
# ─────────────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Parameters ─────────────────────────────────────────────────────────────
  parameters:
...
    ProviderId:
      name: provider_id
      in: path
      required: true
      schema: { type: string, format: uuid }

  # ── Schemas ─────────────────────────────────────────────────────────────────
  schemas:
...
    AiProvider:
      type: object
      properties:
        id:            { type: string, format: uuid }
        provider:      { type: string, enum: [openai, anthropic, gemini, local, custom] }
        display_name:  { type: string, example: "My LM Studio" }
        base_url:      { type: string, example: "http://localhost:1234/v1" }
        default_model: { type: string, example: "gpt-4o" }
        has_key:       { type: boolean, description: "True if an API key is stored — key itself is never returned" }
        is_active:     { type: boolean }
        created_at:    { type: string, format: date-time }
    CreateAiProviderRequest:
      type: object
      required: [provider, base_url]
      properties:
        provider:      { type: string, enum: [openai, anthropic, gemini, local, custom] }
        display_name:  { type: string }
        api_key:       { type: string, description: "Encrypted server-side. Never returned in responses.", nullable: true }
        base_url:      { type: string, example: "https://api.openai.com/v1" }
        default_model: { type: string }
    UpdateAiProviderRequest:
      type: object
      properties:
        display_name:  { type: string }
        api_key:       { type: string, nullable: true, description: "Send to rotate the key" }
        base_url:      { type: string }
        default_model: { type: string }
        is_active:     { type: boolean }
...
    ErrorResponse:
      type: object
      properties:
        message: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }

  # ── Responses ───────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
paths:
...
  /me/ai-providers:
    get:
      tags: [Me]
      summary: List own AI provider connections
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/AiProvider' } }
        '401': { $ref: '#/components/responses/Unauthorized' }

    post:
      tags: [Me]
      summary: Add an AI provider connection
      description: |
        Encrypts `api_key` with Laravel's APP_KEY before storing. The key is
        **never** returned in any response — only the `has_key` boolean.
        A user can have at most one row per `provider` value.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAiProviderRequest' }
      responses:
        '201':
          description: Provider added
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiProvider' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
  /me/ai-providers/{provider_id}:
    parameters:
      - $ref: '#/components/parameters/ProviderId'

    put:
      tags: [Me]
      summary: Update an AI provider connection
      description: |
        Send `api_key` to rotate the encrypted key. Omit it to keep the
        existing one. Foreign IDs return 404 (not 403) to avoid leaking
        which rows exist.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAiProviderRequest' }
      responses:
        '200':
          description: Provider updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiProvider' }
        '404': { $ref: '#/components/responses/NotFound' }

    delete:
      tags: [Me]
      summary: Remove an AI provider connection
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...